식당바텀리스트 이미지 api 연결 완료
- 검색결과 data class 내 restaurant_id's type ; Int -> Long 수정
- 식당바텀리스트 이미지 api 연결 완료
* 식당바텀리스트 이미지 바인딩 필요 !!!
* 식당결과 클릭 시 해당 식당으로 카메라 이동 필요 !!!

+++ app/src/main/java/com/bapool/bapool/retrofit/data/Restaurant.kt
@@ -86,7 +86,7 @@ data class RestaurantSearch(
     val category: String,
     val num_of_party: Int,
     val restaurant_address: String,
-    val restaurant_id: Int,
+    val restaurant_id: Long,
     val restaurant_name: String,
     val restaurant_longitude: Double,
     val restaurant_latitude: Double
